database_restore for cassandra running as a docker container.

+++ scripts/database_restore.sh
@@ -1,7 +1,9 @@
 #!/bin/bash
 
 ## Restore a KEYSPACE and its data from a CQL file with the schema of the
-## KEYSPACE and an tbz archive containing the snapshot
+## KEYSPACE and an tbz archive containing the snapshot.
+
+## Works with cassandra docker container, should be run from host
 
 ## Complete variables before running:
 ## IP: IP of cassandra server
@@ -11,33 +13,46 @@
 ## SNAPSHOT: choose a name for the backup
 ## SNAPSHOT_INDEX: index of the snapshot (1, 20210401 ...)
 
-#IP=127.0.0.1
+IP=127.0.0.1 # check in /etc/cassandra/cassandra.yaml
 TMP=/tmp/cassandra
-#BACKUP_FOLDER=/opt/backups/cassandra
 SOURCE_KEYSPACE="thehive"
 TARGET_KEYSPACE="thehive"
+SNAPSHOT_PATH=""
 SNAPSHOT=""
 SNAPSHOT_INDEX="1"
 DOCKER_CONTAINER=cassandra
+CASSANDRA_PASSWORD=""
+CASSANDRA_USER=""
+
+if [ -z "$SNAPSHOT" ]
+then
+  echo "Need to provide SNAPSHOT."
+  echo "Note the formatting of the SNAPSHOT variable - filename without path and extension."
+  echo "Exiting."
+  exit 1
+fi
 
 ## Uncompress data in TMP folder
 rm -rf ${TMP} && mkdir ${TMP}
-tar jxf "${SNAPSHOT}"_${SNAPSHOT_INDEX}.tbz -C ${TMP}
+tar jxf "${SNAPSHOT_PATH}""${SNAPSHOT}"_${SNAPSHOT_INDEX}.tbz -C ${TMP}
+
+#exit
 
 ## Read Cassandra password
 #echo -n "Cassandra admin password: "
 #read -s CASSANDRA_PASSWORD
 
-## Define new KEYSPACE NAME
-#sed -i "s/${SOURCE_KEYSPACE}/${TARGET_KEYSPACE}/g" schema_${SNAPSHOT}_${SNAPSHOT_INDEX}.cql
-
 ## Restore keyspace
-docker exec -u 0 -it $DOCKER_CONTAINER cqlsh -u cassandra -p "${CASSANDRA_PASSWORD}" cassandra --file schema_"${SNAPSHOT}"_${SNAPSHOT_INDEX}.cql
+docker cp schema_"${SNAPSHOT}"_${SNAPSHOT_INDEX}.cql ${DOCKER_CONTAINER}:/tmp/
+docker exec -u 0 -i $DOCKER_CONTAINER cqlsh -u "${CASSANDRA_USER}" -p "${CASSANDRA_PASSWORD}" --file /tmp/schema_"${SNAPSHOT}"_${SNAPSHOT_INDEX}.cql
 
 ## Restore data
-for TABLE in $(cqlsh -u cassandra -p "${CASSANDRA_PASSWORD}" -e "use ${TARGET_KEYSPACE}; DESC tables ;")
+for TABLE in $(docker exec -u 0 -i $DOCKER_CONTAINER cqlsh -u "${CASSANDRA_USER}" -p "${CASSANDRA_PASSWORD}" -e "use ${TARGET_KEYSPACE}; DESC tables ;")
 do
-docker cp ${TMP}/var/lib/cassandra/data/${SOURCE_KEYSPACE}/"${TABLE}"-*/snapshots/"${SNAPSHOT}"_${SNAPSHOT_INDEX}/* cassandra:/var/lib/cassandra/data/${TARGET_KEYSPACE}/"$TABLE"-*/
+  for FILE in $(ls ${TMP}/var/lib/cassandra/data/${SOURCE_KEYSPACE}/"${TABLE}"-*/snapshots/"${SNAPSHOT}"_${SNAPSHOT_INDEX}/)
+  do
+    docker cp $FILE cassandra:/var/lib/cassandra/data/${TARGET_KEYSPACE}/"$TABLE"-*/
+  done
 done
 
 ## Change ownership
@@ -45,7 +60,7 @@ done
 
 
 ## sstableloader
-for TABLE in $(docker exec -u 0 -it $DOCKER_CONTAINER ls /var/lib/cassandra/data/${TARGET_KEYSPACE})
+for TABLE in $(docker exec -u 0 -i $DOCKER_CONTAINER ls /var/lib/cassandra/data/${TARGET_KEYSPACE})
 do
-    docker exec -u 0 -it $DOCKER_CONTAINER sstableloader /var/lib/cassandra/data/${TARGET_KEYSPACE}/"${TABLE}"
+    docker exec -u 0 -i $DOCKER_CONTAINER sh -c "sstableloader -d "${IP}" /var/lib/cassandra/data/${TARGET_KEYSPACE}/${TABLE}"
 done
